Conflicts:
	c-semantics-plugin/src/main/c-semantics/core-extensions/language/common/conversion.k
	c-semantics/semantics/language/common/memory/reading.k

+++ tests/unit-pass/931110-1.c
@@ -1,7 +1,7 @@
 #include <stdlib.h>
 typedef struct
 {
-  short f:3, g:3, h:10;
+  int f:3, g:3, h:10;
 } small;
 
 struct
